Tests and coverage for the entire SQS module; new `close` method.
Tests the various functions in `SQSStream` ensuring basic operations work as intended.

Also added now is a `close` method which stops the readable stream from continuing to ask for more jobs.

The long poll for new jobs is now done on the next tick to give the opportunity for people to call `close` and not enter into some scary infinite loop while testing.